Don't limit width of DAC data to 16 bits (some platforms may support wider).

+++ inc/platform.h
@@ -292,7 +292,7 @@ int  platform_dac_exists(unsigned id);
 #define DAC_INIT_BAD_OPTION          -3
 
 int platform_dac_init(unsigned id, unsigned bits_per_sample, unsigned options);
-void platform_dac_put_sample(unsigned channel_mask, u16 *data);
+void platform_dac_put_sample(unsigned channel_mask, unsigned *data);
 int  platform_dac_check_timer_id(unsigned id, unsigned timer_id);
 
 // *****************************************************************************
